Even More Continued Work On Curse/Transmute Effects, 8/20/2022
Even More Continued Work On Curse/Transmute Effects. Decided to just rework what I was doing and go with a less elegant, but more easy to understand (for me at least) system of just a separate effect for each transmute instead of the variant thing, it was getting way too confusing and weird for my patience, frustration and confusion made doing it the "better" way not worth it. So hopefully in the end this works just fine, even if it's less fancy, 8/20/2022.