Added Folder For Spell Effect Ideas And Notes, 8/22/2022
Added Folder For Spell Effect Ideas And Notes, mostly for those wondering what I plan to be working on, as well as my own thing of backups for the ideas in simple image form I suppose. Likely won't keep after mod is released on the Git Repo atleast, but who knows, maybe I will for the sake of it, 8/22/2022.